refactor(cli): extract default-run pipeline from main() into src/cli/run.js
Part of #411 (item 1). Zero-behavior-change move, continuing the #409 decomposition.
